requirements.txt
discord.py
pylast
PyNaCl
pyfiglet
flask
requests